--- examples/pipelines/starcoder/components/pii_detection/Dockerfile
@@ -1,19 +1,0 @@
-FROM --platform=linux/amd64 pytorch/pytorch:2.0.1-cuda11.7-cudnn8-devel
-
-## System dependencies
-RUN apt-get update && \
-    apt-get upgrade -y && \
-    apt-get install git -y
-
-# install requirements
-COPY requirements.txt ./
-RUN pip3 install --no-cache-dir -r requirements.txt
-
-# Set the working directory to the compoent folder
-WORKDIR /component/src
-
-# Copy over src-files and spec of the component
-COPY src/ .
-COPY fondant_component.yaml ../
-
-ENTRYPOINT ["python", "main.py"]
