--- examples/pipelines/starcoder/components/pii_detection/fondant_component.yaml
@@ -1,15 +1,0 @@
-name: PII detection
-description: A component that detects and redacts Personal Identifiable Information (PII) from code.
-image: ghcr.io/ml6team/pii_detection:latest
-
-consumes:
-  code:
-    fields:
-      content:
-        type: string
-
-produces:
-  code:
-    fields:
-      content:
-        type: string
