--- examples/pipelines/starcoder/components/pii_detection/requirements.txt
@@ -1,6 +1,0 @@
-fondant==0.1.0
-pyarrow>=7.0
-gcsfs==2023.4.0
-regex==2023.5.5
-detect_secrets==1.4.0
-gibberish_detector==0.1.1
