+++ examples/pipelines/starcoder/components/pii_redaction/fondant_component.yaml
@@ -1,5 +1,5 @@
 name: PII redaction
-description: A component that redacts (removes) Personal Identifiable Information (PII) from code.
+description: A component that detects and redacts Personal Identifiable Information (PII) from code.
 image: ghcr.io/ml6team/pii_redaction:latest
 
 consumes:
@@ -8,4 +8,8 @@ consumes:
       content:
         type: string
 
-args:
+produces:
+  code:
+    fields:
+      content:
+        type: string
