+++ examples/pipelines/starcoder/components/pii_redaction/src/main.py
@@ -1,5 +1,6 @@
-"""A component that redacts Personal Identifiable Information (PII) from code."""
+"""A component that detects and redacts Personal Identifiable Information (PII) in code."""
 
+import json
 import logging
 
 import dask.dataframe as dd
@@ -7,13 +8,16 @@ import dask.dataframe as dd
 from fondant.component import TransformComponent
 from fondant.logger import configure_logging
 
+from pii_detection import scan_pii
+from pii_redaction import redact_pii
+
 configure_logging()
 logger = logging.getLogger(__name__)
 
 
-class RedactPIIComponent(TransformComponent):
+class RemovePIIComponent(TransformComponent):
     """
-    Component that redacts PII from code.
+    Component that detects and redacts PII from code.
     """
 
     def transform(
@@ -28,16 +32,27 @@ class RedactPIIComponent(TransformComponent):
         Returns:
             Dask dataframe
         """
+        # detect PII
         result = dataframe.apply(lambda example: scan_pii(text=example.code_content), axis=1, result_type="expand", meta={0: object, 1: bool, 2: int},)
         result.columns = ["code_secrets", "code_has_secrets", "code_number_secrets"]
-        
+
         dataframe = dataframe.merge(
             result, left_index=True, right_index=True
         )
+
+        # redact PII
+        # we use random replacements by default
+        with open("replacements.json", "r") as f:
+            replacements = json.load(f)
+
+        dataframe['code_content'] = dataframe.apply(lambda example: redact_pii(text=example.code_content, secrets=example.code_secrets, has_secrets=example.code_has_secrets, replacements=replacements), axis=1, meta=(None, 'str'))
+        dataframe = dataframe.drop(["code_secrets", "code_has_secrets", "code_number_secrets"], axis=1)
+
+        print(dataframe.head())
 
         return dataframe
     
 
 if __name__ == "__main__":
-    component = RedactPIIComponent.from_file()
+    component = RemovePIIComponent.from_file()
     component.run()
