+++ examples/pipelines/starcoder/components/pii_redaction/src/pii_redaction.py
