+++ examples/pipelines/starcoder/components/pii_redaction/src/utils/keys_detection.py
